pyproject.toml
[project]
name = "eatool"
version = "1.0.0"
description = "EA Team tool that does things"
authors = [
  { name = "Philipp Kewisch", email = "code@kewis.ch" }
]
requires-python = ">=3.6"
dependencies = [
  "indico-cli @ git+https://github.com/kewisch/indico-cli",
  "click",
  "pyyaml",
  "google-api-python-client",
  "google-auth-httplib2",
  "google-auth-oauthlib"
]
[project.urls]
homepage = "https://github.com/kewisch/canonical-eatool"

[project.optional-dependencies]
dev = [
  "black",
  "isort",
  "flake8",
  "Flake8-pyproject",
  "nox"
]

[project.scripts]
eatool = "eatool.cli:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.flake8]
ignore = ["E203", "E266", "E501", "W503"]
max-line-length = 100
max-complexity = 18
select = ["B","C","E","F","W","T4"]


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100
